Вынес проверки в отдельные классы.
