+++ src/main/java/praktikum/order/OrderClient.java
@@ -1,16 +1,14 @@
-package praktikum.base;
+package praktikum.order;
 
 import io.qameta.allure.Step;
-import io.restassured.RestAssured;
-import io.restassured.http.ContentType;
 import io.restassured.response.Response;
-import org.junit.jupiter.api.BeforeAll;
+import praktikum.base.BaseHttpClient;
 import praktikum.models.Order;
 
 import static io.restassured.RestAssured.given;
 import static org.hamcrest.Matchers.notNullValue;
 
-public class BaseOrderTest {
+public class OrderClient {
 
     private final BaseHttpClient baseHttpClient = new BaseHttpClient();
     @Step("Создание заказа")
@@ -22,13 +20,6 @@ public class BaseOrderTest {
                 .post("/orders");
     }
 
-    @Step("Проверка корректности создания заказа")
-    public void assertSuccessCreateOrder(Response response) {
-        response.then()
-                .statusCode(201)
-                .body("track", notNullValue());
-    }
-
     @Step("Получение списка заказов")
     public Response getOrderList() {
         return given()
@@ -37,11 +28,5 @@ public class BaseOrderTest {
                 .get("/orders");
     }
 
-    @Step("Проверка корректности получения списка заказов")
-    public void assertSuccessGetOrderList(Response response) {
-        response.then()
-                .statusCode(200)
-                .body("orders", notNullValue());
-    }
 
 }
