+++ src/test/java/TestListOrder.java
@@ -1,17 +1,19 @@
 import io.restassured.response.Response;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import praktikum.base.BaseOrderTest;
+import praktikum.order.OrderChecks;
+import praktikum.order.OrderClient;
 
 
 
 public class TestListOrder {
-    private final BaseOrderTest baseOrderTest = new BaseOrderTest();
+    private final OrderClient orderClient = new OrderClient();
+    private final OrderChecks check = new OrderChecks();
 
     @DisplayName("Проверка корректности получения списка заказов")
     @Test
     public void testGetOrderList() {
-        Response response = baseOrderTest.getOrderList();
-        baseOrderTest.assertSuccessGetOrderList(response); // содержащий 30 элементов - значение по умолчанию.
+        Response response = orderClient.getOrderList();
+        check.assertSuccessGetOrderList(response); // содержащий 30 элементов - значение по умолчанию.
     }
 }
